Create Real Estate
Sooooo, right now trying to figure out how (without cracking thanks) to download automatically from google images on an inventory sheet. 

We have the link for the images, but more than making the code work (it does) the problem comes when one link is broken or a cell empties on the Link Row. There it begins to accumulate all the images in the last cell that took the data correctly.